Reorganize backend files into internal packages
