--- backend/routes.go
@@ -1,9 +1,0 @@
-package main
-
-import "net/http"
-
-func setupRoutes() {
-	http.HandleFunc("/api/speedtest", speedTestHandler)
-	http.HandleFunc("/api/server-names", getServerNamesHandler)
-	http.HandleFunc("/api/settings", handleSettings)
-}
